chore(ionic-react): better organize schematics
