+++ libs/ionic-react/src/schematics/application/lib/add-dependencies.ts
@@ -1,0 +1,12 @@
+import { Rule } from '@angular-devkit/schematics';
+import { addDepsToPackageJson } from '@nrwl/workspace';
+import { ionicReactRouterVersion } from '../../../utils/versions';
+
+export function addDependencies(): Rule {
+  return addDepsToPackageJson(
+    {
+      '@ionic/react-router': ionicReactRouterVersion,
+    },
+    {}
+  );
+}
