+++ libs/ionic-react/src/schematics/init/schematic.ts
@@ -1,28 +1,8 @@
 import { chain, noop, Rule } from '@angular-devkit/schematics';
-import { addDepsToPackageJson, addPackageWithInit } from '@nrwl/workspace';
+import { addPackageWithInit } from '@nrwl/workspace';
 import { setDefaultCollection } from '@nrwl/workspace/src/utils/rules/workspace';
-import {
-  ioniconsVersion,
-  ionicReactVersion,
-  nxtendVersion,
-  testingLibraryJestDomVersion,
-  testingLibraryUserEventVersion,
-} from '../../utils/versions';
+import { addDependencies } from './lib/add-dependencies';
 import { InitSchematicSchema } from './schema';
-
-function addDependencies(): Rule {
-  return addDepsToPackageJson(
-    {
-      '@ionic/react': ionicReactVersion,
-      ionicons: ioniconsVersion,
-    },
-    {
-      '@nxtend/ionic-react': nxtendVersion,
-      '@testing-library/user-event': testingLibraryUserEventVersion,
-      '@testing-library/jest-dom': testingLibraryJestDomVersion,
-    }
-  );
-}
 
 export default function (options: InitSchematicSchema): Rule {
   return chain([
